requirements.txt
flask==2.0.1
flask_login==0.5.0
flask_migrate==3.1.0
WTForms==2.3.3
flask_wtf==0.15.1
flask_sqlalchemy==2.5.1
sqlalchemy==1.4.23
email_validator==1.1.3
python-decouple==3.4
gunicorn==20.1.0
jinja2==3.0.1
flask-restx==0.5.1

 
certifi==2020.6.20
chardet==3.0.4 
cmake==3.18.2.post1
decorator==4.4.2
dlib==19.18.0
face-recognition==1.3.0
face-recognition-models==0.3.0
idna==2.10
imageio==2.9.0
imageio-ffmpeg==0.4.2
moviepy==1.0.3
numpy
opencv-python==4.4.0.46
Pillow==8.0.1
proglog==0.1.9
requests==2.24.0
tqdm==4.51.0
urllib3==1.25.11
wincertstore==0.2

bidict==0.22.1
click==8.1.3
dnspython==2.3.0
eventlet==0.33.3 
Flask-SocketIO==5.3.3
greenlet==2.0.2
itsdangerous==2.1.2 
MarkupSafe==2.1.2
python-engineio==4.4.0
python-socketio==5.8.0
six==1.16.0
Werkzeug==2.0.1


ultralytics
deepface
retina-face